requirements.txt
mlflow==3.5.1
ray[tune]==2.51.1
numpy==2.2.6
scipy==1.15.3
ipykernel==7.1.0
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
torchmetrics==1.8.2
streamlit==1.51.0
onnx==1.19.1
nibabel==5.3.2
onnxruntime-gpu==1.23.2
monai[nibabel]
tqdm==4.67.1
# pydantic
# fastapi
# uvicorn[standard]